Fix #355: Handle malformed class name
The binary name encoding for Scala has been broken for a while. It does
not follow the Java source name rules because it does not append a `$`
after the mirror class of an object, which already ends in `$`.

Zinc works around this fix in the only way it can: it gets the enclosing
class, which is obtained from the `InnerClasses` java classfile
metadata, and it reconstructs the path from there. It circumvents, by
design, the core issue: it doesn't call Java's name parser at all.

This issue has been fixed in the JDK9, but unfortunately it's going to
be tormenting us for a while because 2.13 does not target Java 9, and
we're still at 2.12.x.

This is the fix in JDK9:
http://bugs.java.com/bugdatabase/view_bug.do?bug_id=8057919

In order to fully fix this issue, the Scala compiler should do the
desired behaviour in 2.13.x. 2.12.x is already lost since this fix is
binary incompatible. The appropriate JIRA ticket is here:
https://github.com/scala/bug/issues/2034.