Make sure to compile only the bridge's sources
When a compiler bridge has dependencies, we download these dependencies,
along with their sources. Later, when we collect the sources jar, we
used to include these sources and compile them along with the sources
of the compiler bridge.

This situation was less than optimal, because the compilation could take
a much longer time than it should, the bridge JAR would include more
classfiles than what is expected and necessary.

This commit changes the component compiler so that it collects only the
sources of the bridge, and considers all other non-sources jar as
dependencies that should be put on the classpath while compiling the
bridge.

This was discovered while integrating Dottydoc with sbt. Unlike scaladoc
which lives in the same jar as scalac and which has no external
dependencies, dottydoc is a separate project that depends on dotty and
on many more libraries. Because the bridge needs to depend on dottydoc
to be able to invoke it, I noticed that the component compiler would
also download the sources of dottydoc, dotty, etc. and that compiling
the bridge would take a very long time.